Record the time the job spent in the queue
It would be a useful metric for me and my company to track how long jobs
are in the queue, so we can alert on an increase in that timing, letting
use know something is wrong, or we need to add more workers/threads,
etc. This tracks queued_time by subtracting the `start` time from the
job's `created_at`, which should be when it was enqueued.