Extract tag building logic into shareable class
There is a good amount of logic in the server middleware to build the
tags for the metrics we are reporting. This change extracts that logic
into a separate class that we can reuse from other middleware.

This is in preparation for adding metrics to the Client-side middleware
(enqueuing jobs), which also needs most of this tags logic. (added in
next commit)