Track jobs enqueued through a Client middleware
In addition to tracking how many jobs were executed, through the Server
middleware, we also track how many jobs are enqueued, thorugh Client
middleware.

Users of this gem have the option of using either or both middlewares
depending on what they need.

This change is backwards compatible with the existing gem, releasing it
would be a minor version bump.