Track jobs enqueued through a Client middleware (#20)
* Extract tag building logic into shareable class

There is a good amount of logic in the server middleware to build the
tags for the metrics we are reporting. This change extracts that logic
into a separate class that we can reuse from other middleware.

This is in preparation for adding metrics to the Client-side middleware
(enqueuing jobs), which also needs most of this tags logic. (added in
next commit)

* Track jobs enqueued through a Client middleware

In addition to tracking how many jobs were executed, through the Server
middleware, we also track how many jobs are enqueued, thorugh Client
middleware.

Users of this gem have the option of using either or both middlewares
depending on what they need.

This change is backwards compatible with the existing gem, releasing it
would be a minor version bump.