Prepare v0.6.0 release, and support `dogstats-ruby` v4.8 and >= 5.0 (#22)
* Switch from TravisCI to Github Actions

* Pin `dogstatsd-ruby` to v4.*

Our gem doesn't quite work yet with Dogstats v5.*. We'll deal with that
upgrade later.

* Add `rubocop-bsm` gem, update bundle and format everything

* Bump version to 0.6.0

* Support `dogstatsd-ruby` >= 5.0

`dogstatsd-ruby` has major backwards incompatibilities between
v4.8.3 and 5.0.0.

This gem is compatible with both, and CI tests both versions.

This required a number of changes:
- A new gemfile pointing to v4.8, so we can test against either version.
- Specifying a matrix for Github Actions to test against both Gemfiles
- Explicitly flushing when incrementing metrics (more below)
- Changing our Mock of `Datadog::Statsd` to match its new internals

In order to test the stats sent, we are subclassing and overwriting
the internals of the `Datadog::Statsd` class.

These internals changed in 5.0.0: method `send_stats` no longer calls
`send_stat`. Instead it uses an internal forwarder that it would be
very hacky to try to override. Instead, we're not overriding `send_stats`,
which has a different signature, and "converting the message" to the
same parameter we used to receive.

We're also calling `.flush` if the Statsd object responds to that method,
which only exists in 5.0. Since we're instantiating our own instance of
Statsd, I believe this is the correct thing to do, but i'm not 100% sure.